+++ rasa_nlu/classifiers/keyword_intent_classifier.py
@@ -24,7 +24,9 @@ class KeywordIntentClassifier(Component):
     def process(self, message, **kwargs):
         # type: (Message, **Any) -> None
 
-        message.set("intent", {"name": self.parse(message.text), "confidence": 1.0}, add_to_output=True)
+        intent = {"name": self.parse(message.text), "confidence": 1.0}
+        message.set("intent", intent,
+                    add_to_output=True)
 
     def parse(self, text):
         # type: (Text) -> Text