+++ rasa_nlu/persistor.py
@@ -14,26 +14,22 @@ import botocore
 from builtins import object
 from typing import Optional, Tuple, List, Text
 
-from rasa_nlu.config import RasaNLUConfig
+from rasa_nlu.config import RasaNLUModelConfig
 
 logger = logging.getLogger(__name__)
 
 
-def get_persistor(config):
-    # type: (RasaNLUConfig) -> Optional[Persistor]
+def get_persistor(name):
+    # type: (Text) -> Optional[Persistor]
     """Returns an instance of the requested persistor.
 
     Currently, `aws` and `gcs` are supported"""
 
-    if 'storage' not in config:
-        raise KeyError("No persistent storage specified. Supported values "
-                       "are {}".format(", ".join(['aws', 'gcs'])))
-
-    if config['storage'] == 'aws':
-        return AWSPersistor(config['aws_region'], config['bucket_name'],
-                            config['aws_endpoint_url'])
-    elif config['storage'] == 'gcs':
-        return GCSPersistor(config['bucket_name'])
+    if name == 'aws':
+        return AWSPersistor(os.environ.get("BUCKET_NAME"),
+                            os.environ.get("AWS_ENDPOINT_URL"))
+    elif name == 'gcs':
+        return GCSPersistor(os.environ.get("BUCKET_NAME"))
     else:
         return None
 
@@ -41,15 +37,16 @@ def get_persistor(config):
 class Persistor(object):
     """Store models in cloud and fetch them when needed"""
 
-    def persist(self, mode_directory, model_name, project):
+    def persist(self, model_directory, model_name, project):
         # type: (Text) -> None
         """Uploads a model persisted in the `target_dir` to cloud storage."""
 
-        if not os.path.isdir(mode_directory):
+        if not os.path.isdir(model_directory):
             raise ValueError("Target directory '{}' not "
-                             "found.".format(mode_directory))
+                             "found.".format(model_directory))
 
-        file_key, tar_path = self._compress(mode_directory, model_name, project)
+        file_key, tar_path = self._compress(
+                model_directory, model_name, project)
         self._persist_tar(file_key, tar_path)
 
     def retrieve(self, model_name, project, target_path):
@@ -88,11 +85,13 @@ class Persistor(object):
     def _compress(self, model_directory, model_name, project):
         # type: (Text) -> Tuple[Text, Text]
         """Creates a compressed archive and returns key and tar."""
+        import tempfile
 
+        dirpath = tempfile.mkdtemp()
         base_name = self._tar_name(model_name, project, include_extension=False)
         tar_name = shutil.make_archive(base_name, 'gztar',
                                        root_dir=model_directory,
-                                       base_dir=".")
+                                       base_dir=dirpath)
         file_key = os.path.basename(tar_name)
         return file_key, tar_name
 
@@ -100,11 +99,12 @@ class Persistor(object):
     def _project_prefix(project):
         # type: (Text) -> Text
 
-        return '{}___'.format(project or RasaNLUConfig.DEFAULT_PROJECT_NAME)
+        p = project or RasaNLUModelConfig.DEFAULT_PROJECT_NAME
+        return '{}___'.format(p)
 
     @staticmethod
     def _project_and_model_from_filename(filename):
-        # type: (Text) -> Text
+        # type: (Text) -> (Text, Text)
 
         split = filename.split("___")
         if len(split) > 1:
@@ -126,7 +126,8 @@ class Persistor(object):
         # type: (Text, Text) -> None
 
         with tarfile.open(compressed_path, "r:gz") as tar:
-            tar.extractall(target_path)  # project dir will be created if it not exists
+            tar.extractall(
+                target_path)  # project dir will be created if it not exists
 
 
 class AWSPersistor(Persistor):
@@ -134,13 +135,11 @@ class AWSPersistor(Persistor):
 
     Fetches them when needed, instead of storing them on the local disk."""
 
-    def __init__(self, aws_region, bucket_name, endpoint_url=None):
-        # type: (Text, Text, Optional[Text]) -> None
+    def __init__(self, bucket_name, endpoint_url=None):
+        # type: (Text, Optional[Text]) -> None
         super(AWSPersistor, self).__init__()
-        self.s3 = boto3.resource('s3',
-                                 region_name=aws_region,
-                                 endpoint_url=endpoint_url)
-        self._ensure_bucket_exists(bucket_name, aws_region)
+        self.s3 = boto3.resource('s3', endpoint_url=endpoint_url)
+        self._ensure_bucket_exists(bucket_name)
         self.bucket_name = bucket_name
         self.bucket = self.s3.Bucket(bucket_name)
 
@@ -167,8 +166,9 @@ class AWSPersistor(Persistor):
                                                  self.aws_region))
             return []
 
-    def _ensure_bucket_exists(self, bucket_name, aws_region):
-        bucket_config = {'LocationConstraint': aws_region}
+    def _ensure_bucket_exists(self, bucket_name):
+        bucket_config = {
+            'LocationConstraint': boto3.DEFAULT_SESSION.region_name}
         try:
             self.s3.create_bucket(Bucket=bucket_name,
                                   CreateBucketConfiguration=bucket_config)