+++ rasa_nlu/project.py
@@ -15,7 +15,8 @@ from threading import Lock
 from rasa_nlu import utils
 from typing import Text, List
 
-from rasa_nlu.config import RasaNLUConfig
+from rasa_nlu.classifiers.keyword_intent_classifier import \
+    KeywordIntentClassifier
 from rasa_nlu.model import Metadata, Interpreter
 
 logger = logging.getLogger(__name__)
@@ -26,8 +27,11 @@ FALLBACK_MODEL_NAME = "fallback"
 
 
 class Project(object):
-    def __init__(self, config=None, component_builder=None, project=None):
-        self._config = config
+    def __init__(self,
+                 component_builder=None,
+                 project=None,
+                 project_dir=None,
+                 remote_storage=None):
         self._component_builder = component_builder
         self._models = {}
         self.status = 0
@@ -37,9 +41,10 @@ class Project(object):
         self._readers_count = 0
         self._path = None
         self._project = project
+        self.remote_storage = remote_storage
 
-        if project:
-            self._path = os.path.join(self._config['path'], project)
+        if project and project_dir:
+            self._path = os.path.join(project_dir, project)
         self._search_for_models()
 
     def _begin_read(self):
@@ -59,9 +64,10 @@ class Project(object):
 
     def _load_local_model(self, requested_model_name=None):
         if requested_model_name is None:  # user want latest model
-            # NOTE: for better parse performance, currently although user may want
-            # latest model by set requested_model_name explicitly to None, we are
-            # not refresh model list from local and cloud which is pretty slow.
+            # NOTE: for better parse performance, currently although
+            # user may want latest model by set requested_model_name
+            # explicitly to None, we are not refresh model list
+            # from local and cloud which is pretty slow.
             # User can specific requested_model_name to the latest model name,
             # then model will be cached, this is a kind of workaround to
             # refresh latest project model.
@@ -109,7 +115,8 @@ class Project(object):
         self._loader_lock.acquire()
         try:
             if not self._models.get(model_name):
-                self._models[model_name] = self._interpreter_for_model(model_name)
+                interpreter = self._interpreter_for_model(model_name)
+                self._models[model_name] = interpreter
         finally:
             self._loader_lock.release()
 
@@ -127,9 +134,12 @@ class Project(object):
 
     def unload(self, model_name):
         self._writer_lock.acquire()
-        unloaded_model = self._models.pop(model_name)
-        self._writer_lock.release()
-        return unloaded_model
+        try:
+            del self._models[model_name]
+            self._models[model_name] = None
+            return model_name
+        finally:
+            self._writer_lock.release()
 
     def _latest_project_model(self):
         """Retrieves the latest trained model for an project"""
@@ -145,12 +155,15 @@ class Project(object):
             return FALLBACK_MODEL_NAME
 
     def _fallback_model(self):
-        meta = Metadata({"pipeline": ["intent_classifier_keyword"]}, "")
-        return Interpreter.create(meta, self._config, self._component_builder)
+        meta = Metadata({"pipeline": [{
+            "name": "intent_classifier_keyword",
+            "class": utils.module_path_from_object(KeywordIntentClassifier())
+        }]}, "")
+        return Interpreter.create(meta, self._component_builder)
 
     def _search_for_models(self):
         model_names = (self._list_models_in_dir(self._path) +
-                       self._list_models_in_cloud(self._config))
+                       self._list_models_in_cloud())
         if not model_names:
             if FALLBACK_MODEL_NAME not in self._models:
                 self._models[FALLBACK_MODEL_NAME] = self._fallback_model()
@@ -161,8 +174,7 @@ class Project(object):
 
     def _interpreter_for_model(self, model_name):
         metadata = self._read_model_metadata(model_name)
-        return Interpreter.create(metadata, self._config,
-                                  self._component_builder)
+        return Interpreter.create(metadata, self._component_builder)
 
     def _read_model_metadata(self, model_name):
         if model_name is None:
@@ -176,20 +188,28 @@ class Project(object):
 
             # download model from cloud storage if needed and possible
             if not os.path.isdir(path):
-                self._load_model_from_cloud(model_name, path, self._config)
+                self._load_model_from_cloud(model_name, path)
 
             return Metadata.load(path)
 
     def as_dict(self):
         return {'status': 'training' if self.status else 'ready',
-                'available_models': list(self._models.keys())}
+                'available_models': list(self._models.keys()),
+                'loaded_models': self._list_loaded_models()}
 
-    def _list_models_in_cloud(self, config):
-        # type: (RasaNLUConfig) -> List[Text]
+    def _list_loaded_models(self):
+        models = []
+        for model, interpreter in self._models.items():
+            if interpreter is not None:
+                models.append(model)
+        return models
+
+    def _list_models_in_cloud(self):
+        # type: () -> List[Text]
 
         try:
             from rasa_nlu.persistor import get_persistor
-            p = get_persistor(config)
+            p = get_persistor(self.remote_storage)
             if p is not None:
                 return p.list_models(self._project)
             else:
@@ -199,10 +219,10 @@ class Project(object):
                         "{}".format(self._project, e))
             return []
 
-    def _load_model_from_cloud(self, model_name, target_path, config):
+    def _load_model_from_cloud(self, model_name, target_path):
         try:
             from rasa_nlu.persistor import get_persistor
-            p = get_persistor(config)
+            p = get_persistor(self.remote_storage)
             if p is not None:
                 p.retrieve(model_name, self._project, target_path)
             else: