+++ rasa_nlu/registry.py
@@ -3,58 +3,50 @@ and preconfigured templates.
 
 Hence, it imports all of the components. To avoid cycles, no component should
 import this in module scope."""
-from __future__ import unicode_literals
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import typing
 from rasa_nlu import utils
-from typing import Any
-from typing import Dict
-from typing import List
-from typing import Optional
-from typing import Text
-from typing import Type
-
 from rasa_nlu.classifiers.keyword_intent_classifier import \
     KeywordIntentClassifier
-from rasa_nlu.classifiers.mitie_intent_classifier import MitieIntentClassifier
 from rasa_nlu.classifiers.sklearn_intent_classifier import \
     SklearnIntentClassifier
 from rasa_nlu.classifiers.tflearn_intent_classifier import \
     TflearnIntentClassifier
+from rasa_nlu.extractors.crf_entity_extractor import CRFEntityExtractor
 from rasa_nlu.extractors.duckling_extractor import DucklingExtractor
 from rasa_nlu.extractors.duckling_http_extractor import DucklingHTTPExtractor
 from rasa_nlu.extractors.entity_synonyms import EntitySynonymMapper
-from rasa_nlu.extractors.mitie_entity_extractor import MitieEntityExtractor
 from rasa_nlu.extractors.spacy_entity_extractor import SpacyEntityExtractor
-from rasa_nlu.extractors.crf_entity_extractor import CRFEntityExtractor
-from rasa_nlu.featurizers.mitie_featurizer import MitieFeaturizer
 from rasa_nlu.featurizers.ngram_featurizer import NGramFeaturizer
 from rasa_nlu.featurizers.regex_featurizer import RegexFeaturizer
 from rasa_nlu.featurizers.spacy_featurizer import SpacyFeaturizer
 from rasa_nlu.model import Metadata
-from rasa_nlu.tokenizers.mitie_tokenizer import MitieTokenizer
 from rasa_nlu.tokenizers.spacy_tokenizer import SpacyTokenizer
 from rasa_nlu.tokenizers.whitespace_tokenizer import WhitespaceTokenizer
-from rasa_nlu.utils.mitie_utils import MitieNLP
 from rasa_nlu.utils.spacy_utils import SpacyNLP
+from typing import Any
+from typing import Optional
+from typing import Text
+from typing import Type
 
 if typing.TYPE_CHECKING:
     from rasa_nlu.components import Component
-    from rasa_nlu.config import RasaNLUConfig
+    from rasa_nlu.config import RasaNLUModelConfig, RasaNLUModelConfig
 
 # Classes of all known components. If a new component should be added,
 # its class name should be listed here.
 component_classes = [
-    SpacyNLP, MitieNLP,
-    SpacyEntityExtractor, MitieEntityExtractor, DucklingExtractor,
+    SpacyNLP,
+    SpacyEntityExtractor, DucklingExtractor,
     CRFEntityExtractor, DucklingHTTPExtractor,
     EntitySynonymMapper,
-    SpacyFeaturizer, MitieFeaturizer, NGramFeaturizer, RegexFeaturizer,
-    MitieTokenizer, SpacyTokenizer, WhitespaceTokenizer,
-    SklearnIntentClassifier, TflearnIntentClassifier, MitieIntentClassifier, KeywordIntentClassifier,
+    SpacyFeaturizer, NGramFeaturizer, RegexFeaturizer,
+    SpacyTokenizer, WhitespaceTokenizer,
+    SklearnIntentClassifier, TflearnIntentClassifier, KeywordIntentClassifier,
 ]
 
 # Mapping from a components name to its class to allow name based lookup.
@@ -80,23 +72,6 @@ registered_pipeline_templates = {
         "ner_synonyms",
         "intent_classifier_tflearn",
     ],
-    "mitie": [
-        "nlp_mitie",
-        "tokenizer_mitie",
-        "ner_mitie",
-        "ner_synonyms",
-        "intent_entity_featurizer_regex",
-        "intent_classifier_mitie",
-    ],
-    "mitie_sklearn": [
-        "nlp_mitie",
-        "tokenizer_mitie",
-        "ner_mitie",
-        "ner_synonyms",
-        "intent_entity_featurizer_regex",
-        "intent_featurizer_mitie",
-        "intent_classifier_sklearn",
-    ],
     "keyword": [
         "intent_classifier_keyword",
     ],
@@ -106,26 +81,33 @@ registered_pipeline_templates = {
     # generate this automatically.
     "all_components": [
         "nlp_spacy",
-        "nlp_mitie",
         "tokenizer_whitespace",
-        "tokenizer_mitie",
         "tokenizer_spacy",
-        "intent_featurizer_mitie",
         "intent_featurizer_spacy",
         "intent_featurizer_ngrams",
         "intent_entity_featurizer_regex",
-        "ner_mitie",
         "ner_crf",
         "ner_spacy",
         "ner_duckling",
         "ner_duckling_http",
         "ner_synonyms",
         "intent_classifier_keyword",
-        "intent_classifier_sklearn",
         "intent_classifier_tflearn",
-        "intent_classifier_mitie",
+        "intent_classifier_sklearn",
     ]
 }
+
+
+def pipeline_template(s):
+    components = registered_pipeline_templates.get(s)
+
+    if components:
+        # converts the list of components in the configuration
+        # format expected (one json object per component)
+        return [{"name": c} for c in components]
+
+    else:
+        return None
 
 
 def get_component_class(component_name):
@@ -162,7 +144,7 @@ def load_component_by_name(component_name,  # type: Text
 
 
 def create_component_by_name(component_name, config):
-    # type: (Text, RasaNLUConfig) -> Optional[Component]
+    # type: (Text, RasaNLUModelConfig) -> Optional[Component]
     """Resolves a component and calls it's create method to init it based on a
     previously persisted model."""