+++ rasa_nlu/tokenizers/spacy_tokenizer.py
@@ -1,23 +1,18 @@
-from __future__ import unicode_literals
-from __future__ import print_function
-from __future__ import division
 from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import typing
-from typing import Any
-from typing import Dict
-from typing import List
-from typing import Optional
-from typing import Text
+from typing import Any, List
 
-from rasa_nlu.config import RasaNLUConfig
-from rasa_nlu.tokenizers import Tokenizer, Token
 from rasa_nlu.components import Component
+from rasa_nlu.config import RasaNLUModelConfig
+from rasa_nlu.tokenizers import Tokenizer, Token
 from rasa_nlu.training_data import Message
 from rasa_nlu.training_data import TrainingData
 
 if typing.TYPE_CHECKING:
-    from spacy.language import Language
     from spacy.tokens.doc import Doc
 
 
@@ -27,7 +22,7 @@ class SpacyTokenizer(Tokenizer, Component):
     provides = ["tokens"]
 
     def train(self, training_data, config, **kwargs):
-        # type: (TrainingData, RasaNLUConfig, **Any) -> None
+        # type: (TrainingData, RasaNLUModelConfig, **Any) -> None
 
         for example in training_data.training_examples:
             example.set("tokens", self.tokenize(example.get("spacy_doc")))