+++ tests/base/test_featurizers.py
@@ -9,13 +9,13 @@ import numpy as np
 import pytest
 
 from rasa_nlu import training_data
-from rasa_nlu.tokenizers.mitie_tokenizer import MitieTokenizer
 from rasa_nlu.tokenizers.spacy_tokenizer import SpacyTokenizer
 from rasa_nlu.training_data import Message
 
 
 @pytest.mark.parametrize("sentence, expected", [
-    ("hey how are you today", [-0.19649599, 0.32493639, -0.37408298, -0.10622784, 0.062756])
+    ("hey how are you today", [-0.19649599, 0.32493639,
+                               -0.37408298, -0.10622784, 0.062756])
 ])
 def test_spacy_featurizer(sentence, expected, spacy_nlp):
     from rasa_nlu.featurizers import spacy_featurizer
@@ -25,40 +25,31 @@ def test_spacy_featurizer(sentence, expected, spacy_nlp):
     assert np.allclose(vecs, doc.vector, atol=1e-5)
 
 
-def test_mitie_featurizer(mitie_feature_extractor, default_config):
-    from rasa_nlu.featurizers.mitie_featurizer import MitieFeaturizer
-
-    default_config["mitie_file"] = os.environ.get('MITIE_FILE')
-    if not default_config["mitie_file"] or not os.path.isfile(default_config["mitie_file"]):
-        default_config["mitie_file"] = os.path.join("data", "total_word_feature_extractor.dat")
-
-    ftr = MitieFeaturizer.load()
-    sentence = "Hey how are you today"
-    tokens = MitieTokenizer().tokenize(sentence)
-    vecs = ftr.features_for_tokens(tokens, mitie_feature_extractor)
-    assert np.allclose(vecs[:5], np.array([0., -4.4551446, 0.26073121, -1.46632245, -1.84205751]), atol=1e-5)
-
-
 def test_ngram_featurizer(spacy_nlp):
     from rasa_nlu.featurizers.ngram_featurizer import NGramFeaturizer
-    ftr = NGramFeaturizer()
-    repetition_factor = 5  # ensures that during random sampling of the ngram CV we don't end up with a one-class-split
+    ftr = NGramFeaturizer({"max_number_of_ngrams": 10})
+
+    # ensures that during random sampling of the ngram CV we don't end up
+    # with a one-class-split
+    repetition_factor = 5
+
+    greet = {"intent": "greet", "text_features": [0.5]}
+    goodbye = {"intent": "goodbye", "text_features": [0.5]}
     labeled_sentences = [
-                            Message("heyheyheyhey", {"intent": "greet", "text_features": [0.5]}),
-                            Message("howdyheyhowdy", {"intent": "greet", "text_features": [0.5]}),
-                            Message("heyhey howdyheyhowdy", {"intent": "greet", "text_features": [0.5]}),
-                            Message("howdyheyhowdy heyhey", {"intent": "greet", "text_features": [0.5]}),
-                            Message("astalavistasista", {"intent": "goodby", "text_features": [0.5]}),
-                            Message("astalavistasista sistala", {"intent": "goodby", "text_features": [0.5]}),
-                            Message("sistala astalavistasista", {"intent": "goodby", "text_features": [0.5]}),
+                            Message("heyheyheyhey", greet),
+                            Message("howdyheyhowdy", greet),
+                            Message("heyhey howdyheyhowdy", greet),
+                            Message("howdyheyhowdy heyhey", greet),
+                            Message("astalavistasista", goodbye),
+                            Message("astalavistasista sistala", goodbye),
+                            Message("sistala astalavistasista", goodbye),
                         ] * repetition_factor
 
     for m in labeled_sentences:
         m.set("spacy_doc", spacy_nlp(m.text))
 
     ftr.min_intent_examples_for_ngram_classification = 2
-    ftr.train_on_sentences(labeled_sentences,
-                           max_number_of_ngrams=10)
+    ftr.train_on_sentences(labeled_sentences)
     assert len(ftr.all_ngrams) > 0
     assert ftr.best_num_ngrams > 0
 
@@ -75,7 +66,7 @@ def test_regex_featurizer(sentence, expected, labeled_tokens, spacy_nlp):
         {"pattern": '[0-9]+', "name": "number", "usage": "intent"},
         {"pattern": '\\bhey*', "name": "hello", "usage": "intent"}
     ]
-    ftr = RegexFeaturizer(patterns)
+    ftr = RegexFeaturizer(known_patterns=patterns)
 
     # adds tokens to the message
     tokenizer = SpacyTokenizer()
@@ -85,12 +76,15 @@ def test_regex_featurizer(sentence, expected, labeled_tokens, spacy_nlp):
 
     result = ftr.features_for_patterns(message)
     assert np.allclose(result, expected, atol=1e-10)
-    assert len(message.get("tokens", [])) > 0  # the tokenizer should have added tokens
+
+    # the tokenizer should have added tokens
+    assert len(message.get("tokens", [])) > 0
     for i, token in enumerate(message.get("tokens")):
         if i in labeled_tokens:
             assert token.get("pattern") in [0, 1]
         else:
-            assert token.get("pattern") is None  # if the token is not part of a regex the pattern should not be set
+            # if the token is not part of a regex the pattern should not be set
+            assert token.get("pattern") is None
 
 
 def test_spacy_featurizer_casing(spacy_nlp):