+++ tests/base/test_multitenancy.py
@@ -10,7 +10,9 @@ import tempfile
 import pytest
 from treq.testing import StubTreq
 
-from rasa_nlu.config import RasaNLUConfig
+from rasa_nlu import config
+from rasa_nlu.config import RasaNLUModelConfig
+from rasa_nlu.data_router import DataRouter
 from rasa_nlu.model import Trainer
 from rasa_nlu.server import RasaNLU
 from tests.utilities import ResponseTest
@@ -18,8 +20,8 @@ from tests.utilities import ResponseTest
 
 @pytest.fixture(scope="module")
 def app(component_builder):
-    """
-    This fixture makes use of the IResource interface of the Klein application to mock Rasa HTTP server.
+    """Use IResource interface of Klein to mock Rasa HTTP server.
+
     :param component_builder:
     :return:
     """
@@ -30,31 +32,17 @@ def app(component_builder):
         root_dir = os.getcwd()
 
     _, nlu_log_file = tempfile.mkstemp(suffix="_rasa_nlu_logs.json")
-    _config = {
-        'write': nlu_log_file,
-        'port': -1,  # unused in test app
-        "pipeline": "keyword",
 
-        "path": os.path.join(root_dir, "test_projects"),
-        "data": os.path.join(root_dir, "data/demo-restaurants.json"),
-        "max_training_processes": 1
-    }
-    train_models(component_builder)
+    train_models(component_builder,
+                 os.path.join(root_dir, "data/examples/rasa/demo-rasa.json"))
 
-    config = RasaNLUConfig(cmdline_args=_config)
-    rasa = RasaNLU(config, component_builder, True)
+    router = DataRouter(os.path.join(root_dir, "test_projects"))
+    rasa = RasaNLU(router, logfile=nlu_log_file, testing=True)
+
     return StubTreq(rasa.app.resource())
 
 
 @pytest.mark.parametrize("response_test", [
-    ResponseTest(
-            "http://dummy-uri/parse?q=food&project=test_project_mitie",
-            {"entities": [], "intent": "affirm", "text": "food"}
-    ),
-    ResponseTest(
-            "http://dummy-uri/parse?q=food&project=test_project_mitie_sklearn",
-            {"entities": [], "intent": "restaurant_search", "text": "food"}
-    ),
     ResponseTest(
             "http://dummy-uri/parse?q=food&project=test_project_spacy_sklearn",
             {"entities": [], "intent": "restaurant_search", "text": "food"}
@@ -90,16 +78,6 @@ def test_get_parse_invalid_model(app, response_test):
 @pytest.mark.parametrize("response_test", [
     ResponseTest(
             "http://dummy-uri/parse",
-            {"entities": [], "intent": "affirm", "text": "food"},
-            payload={"q": "food", "project": "test_project_mitie"}
-    ),
-    ResponseTest(
-            "http://dummy-uri/parse",
-            {"entities": [], "intent": "restaurant_search", "text": "food"},
-            payload={"q": "food", "project": "test_project_mitie_sklearn"}
-    ),
-    ResponseTest(
-            "http://dummy-uri/parse",
             {"entities": [], "intent": "restaurant_search", "text": "food"},
             payload={"q": "food", "project": "test_project_spacy_sklearn"}
     ),
@@ -116,11 +94,16 @@ def test_post_parse(app, response_test):
 def test_post_parse_specific_model(app):
     status = yield app.get("http://dummy-uri/status")
     sjs = yield status.json()
-    model = sjs["available_projects"]["test_project_mitie"]["available_models"][0]
-    query = ResponseTest("http://dummy-uri/parse", {"entities": [], "intent": "affirm", "text": "food"},
-                         payload={"q": "food", "project": "test_project_mitie", "model": model})
+    project = sjs["available_projects"]["test_project_spacy_sklearn"]
+    model = project["available_models"][0]
+    query = ResponseTest("http://dummy-uri/parse",
+                         {"entities": [], "intent": "affirm", "text": "food"},
+                         payload={"q": "food",
+                                  "project": "test_project_spacy_sklearn",
+                                  "model": model})
     response = yield app.post(query.endpoint, json=query.payload)
     assert response.code == 200
+    assert model in project["loaded_models"]
 
 
 @pytest.mark.parametrize("response_test", [
@@ -143,20 +126,17 @@ def test_post_parse_invalid_model(app, response_test):
     assert rjs.get("error").startswith(response_test.expected_response["error"])
 
 
-def train_models(component_builder):
+def train_models(component_builder, data):
     # Retrain different multitenancy models
     def train(cfg_name, project_name):
         from rasa_nlu.train import create_persistor
         from rasa_nlu import training_data
 
-        config = RasaNLUConfig(cfg_name)
-        trainer = Trainer(config, component_builder)
-        training_data = training_data.load_data(config['data'])
+        cfg = config.load(cfg_name)
+        trainer = Trainer(cfg, component_builder)
+        training_data = training_data.load_data(data)
 
         trainer.train(training_data)
-        persistor = create_persistor(config)
-        trainer.persist("test_projects", persistor, project_name)
+        trainer.persist("test_projects", project_name=project_name)
 
-    train("sample_configs/config_mitie.json", "test_project_mitie")
-    train("sample_configs/config_spacy.json", "test_project_spacy_sklearn")
-    train("sample_configs/config_mitie_sklearn.json", "test_project_mitie_sklearn")
+    train("sample_configs/config_spacy.yml", "test_project_spacy_sklearn")