+++ CHANGELOG.rst
@@ -17,10 +17,13 @@ Added
 - support for markdown files containing regex-features or synonyms only
 - added ability to list projects in cloud storage services for model loading
 - server evaluation endpoint at ``POST /evaluate``
-
+- CRF entity recognizer now returns a confidence score when extracting entities
+- server endpoint at ``DELETE /models`` to unload models from server memory
 
 Changed
 -------
+- model configuration is separated from server / train configuration. This is a
+  **breaking change** and models need to be retrained. See migrations guide.
 - Regex features are now sorted internally.
   **retrain your model if you use regex features**
 - The keyword intent classifier now returns ``null`` instead
@@ -30,9 +33,11 @@ Changed
 - The ``CRFEntityExtractor`` now only trains entity examples that have
   ``"extractor": "ner_crf"`` or no extractor at all
 - Ignore hidden files when listing projects or models
+- Docker Images now run on python 3.6 for better non-latin character set support
 
-Fixed
------
+Removed
+-------
+- MITIE support - backend is no longer supported.
 
 
 [0.11.1] - 2018-02-02